[refactor] Extract FlinkWriteSink class
